+++ src/L2Data2Code/L2Data2Code.BaseGenerator/Entities/Constants.cs
@@ -42,8 +42,9 @@ namespace L2Data2Code.BaseGenerator.Entities
         {
             public const string ReferenceTo = "=";
             public const string Collection = "*";
-
         }
+
+        public static string[] NullableTypes = { "bool", "char", "sbyte", "byte", "short", "ushort", "int", "uint", "long", "ulong", "float", "double", "decimal", "DateTime", "TimeSpan" };
 
     }